Changed err testing files
And tests based on them so that utilities across versions
(clang-tidy both 6.0.0 and 8.0.0) will produce the same output,
regardless of platform

+++ tests/conftest.py
@@ -34,7 +34,8 @@ def pytest_exception_interact(node, call, report):
     """See https://docs.pytest.org/en/latest/reference.html#_pytest.hookspec.pytest_exception_interact"""  # noqa: E501
     if report.failed:
         # Clean up temp dirs in tests/files if a test failed.
-        shutil.rmtree("tests/files/temp")
+        if os.path.exists("tests/files/temp"):
+            shutil.rmtree("tests/files/temp")
         # Delete generated files
         for filename in ["ok.plist", "err.plist", "defaults.cfg"]:
             if os.path.exists(filename):
